feat: support displaying proper timestamps
Discord's RPC client has official timestamp support for displaying "elapsed" and "remaining" times. These update every second, as opposed to using the status lines, which will only update every minute or so for everybody but the host.

The old properties have been left in as to not break existing setups.